refactor: refactor router to separate file
