+++ react-project/src/main.tsx
@@ -1,32 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import App from './App';
-import { Aside } from './components/Aside';
-import { ErrorPage } from './components/ErrorPage';
-import { Main } from './components/Main';
+import { RouterProvider } from 'react-router-dom';
 import { StoreProvider } from './context/StoreContext';
 import './index.css';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    errorElement: <ErrorPage />,
-    children: [
-      {
-        path: '/',
-        element: <Main />,
-        children: [
-          {
-            path: '/details/:id',
-            element: <Aside />,
-          },
-        ],
-      },
-    ],
-  },
-]);
+import { router } from './router/router';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
