+++ react-project/src/router/router.tsx
@@ -1,0 +1,27 @@
+import { createBrowserRouter } from 'react-router-dom';
+import App from '../App';
+import { Aside } from '../components/Aside';
+import { ErrorPage } from '../components/ErrorPage';
+import { Main } from '../components/Main';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: '/',
+        element: <Main />,
+        children: [
+          {
+            path: '/details/:id',
+            element: <Aside />,
+          },
+        ],
+      },
+    ],
+  },
+]);
+
+export { router };
